# feat: implement unified Command protocol system to eliminate duplicate analysis execution