… src/config/dependencies.ts
… src/core/unified-command-executor.ts
… src/types/command-protocol.ts